Added example videos of the ER
Fluorescence microscopy videos of the ER. The videos have a framerate of 33.33 frames per second and an exposure time of 30 ms. The size of each pixel is 0.107 microns squared. These videos have been processed using a Fourier bandpass filter with a high-pass filter size of 10 pixels and a low-pass filter of 1 pixel and then using a background subtraction with a 50-pixel rolling ball radius.